Added Requests Model
Requests model tracked pending requests and helps bridge the gap that
the has_friendship gem left. When a pending request is present, a User
sees this on their navbar and can then navigate to the requests index to
accept this request. Need to add declining the request as an option.

Removed the Mailboxer message that automatically sent upon friend
requests. This turned out to be unneccessary and cluttered your mailbox.